refactor: combine base and hermeum-config usecases into mixin (#56)
