+++ apps/dashboard/src/server/usecases/chat.ts
@@ -3,8 +3,7 @@ import { z } from "zod";
 
 import { AgentInput, AgentInputObjectSchema } from "@/entities";
 
-import { FilesUseCase } from "./base";
-import { HermeumConfigLoadable } from "./hermeum-config";
+import { FilesUseCase, HermeumConfigLoadable } from "./mixin";
 
 const DOCS_PATH = "./docs/agent-config";
